+++ src/lib/cron/core-v2/types.ts
@@ -1,0 +1,10 @@
+type MinutelyAlias = "@minutely";
+
+export type CronOptions = {
+  name: string;
+  schedule: Bun.CronWithAutocomplete | MinutelyAlias | (string & {});
+  handler: () => Promise<unknown>;
+  onError?: () => void;
+};
+
+export type CronStatus = "idle" | "running";
